backend/requirements.txt

fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
SQLAlchemy>=1.4.0,<2.0.0
pydantic>=1.8.0,<2.0.0
python-dotenv>=0.19.0,<0.20.0
orjson>=3.6.0,<4.0.0
httpx>=0.23.0,<0.24.0
ortools>=9.5.0,<10.0.0
scikit-learn>=1.5.0,<2.0.0
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
protobuf>=5.26.1,<6.0.0dev
